kubepkg: Refactor common non-command vars and functions into pkg/kubepkg
