+++ BUILD.bazel
@@ -27,13 +27,14 @@ filegroup(
     name = "all-srcs",
     srcs = [
         ":package-srcs",
-        "//build/debs:all-srcs",
         "//cmd/blocking-testgrid-tests:all-srcs",
         "//cmd/krel:all-srcs",
+        "//cmd/kubepkg:all-srcs",
         "//cmd/release-notes:all-srcs",
         "//lib:all-srcs",
         "//pkg/command:all-srcs",
         "//pkg/git:all-srcs",
+        "//pkg/kubepkg:all-srcs",
         "//pkg/notes:all-srcs",
         "//pkg/release:all-srcs",
         "//pkg/util:all-srcs",
