+++ go.mod
@@ -7,6 +7,7 @@ require (
 	github.com/blang/semver v3.5.1+incompatible
 	github.com/gogo/protobuf v1.2.2-0.20190723190241-65acae22fc9d // indirect
 	github.com/golangci/golangci-lint v1.21.0
+	github.com/google/go-github v17.0.0+incompatible
 	github.com/google/go-github/v28 v28.1.1
 	github.com/maxbrunsfeld/counterfeiter/v6 v6.2.2
 	github.com/nozzle/throttler v0.0.0-20180817012639-2ea982251481
@@ -19,6 +20,5 @@ require (
 	google.golang.org/appengine v1.6.1 // indirect
 	google.golang.org/genproto v0.0.0-20190502173448-54afdca5d873 // indirect
 	gopkg.in/src-d/go-git.v4 v4.13.1
-	k8s.io/release/build/debs v0.0.0-20191011003919-ca0d58d1459d
 	k8s.io/test-infra v0.0.0-20190829230513-7ef687d80d22
 )
