+++ pkg/kubepkg/kubepkg_test.go
@@ -14,12 +14,14 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package main
+package kubepkg_test
 
 import (
 	"reflect"
 	"strings"
 	"testing"
+
+	kpkg "k8s.io/release/pkg/kubepkg"
 )
 
 func TestGetKubeadmDependencies(t *testing.T) {
@@ -76,8 +78,8 @@ func TestGetKubeadmDependencies(t *testing.T) {
 
 	for _, tc := range testcases {
 		t.Run(tc.name, func(t *testing.T) {
-			packageDef := packageDefinition{Version: tc.version}
-			deps, err := getKubeadmDependencies(packageDef)
+			packageDef := kpkg.PackageDefinition{Version: tc.version}
+			deps, err := kpkg.GetKubeadmDependencies(packageDef)
 			if err != nil {
 				t.Fatalf("did not expect an error: %v", err)
 			}
